+++ api/audit-trail-gw/src/database/constants.ts
@@ -2,5 +2,6 @@ export const enum CollectionNames {
 	channelInfo = 'channel-info',
 	channelData = 'channel-data',
 	subscriptions = 'subscriptions',
+	subscriptionStates = 'subscription-states',
 	concurrencyLocks = 'concurrency-locks'
 }