+++ api/audit-trail-gw/src/routes/subscription/tests/authorize-subscription.test.ts
@@ -20,7 +20,6 @@ describe('test authorize subscription route', () => {
 		id: 'did:iota:1234',
 		isAuthorized: false,
 		publicKey: 'testpublickey',
-		state: 'teststate',
 		subscriptionLink: 'testlink',
 		type: SubscriptionType.Subscriber
 	};
@@ -117,8 +116,7 @@ describe('test authorize subscription route', () => {
 			channelAddress: 'testaddress',
 			type: SubscriptionType.Subscriber, // caller is not the valid author
 			id: authorId,
-			isAuthorized: true,
-			state: ''
+			isAuthorized: true
 		};
 		const isAuthor = false;
 		jest.spyOn(subscriptionService, 'isAuthor').mockImplementation(async () => isAuthor);
@@ -147,11 +145,11 @@ describe('test authorize subscription route', () => {
 			channelAddress: 'testaddress',
 			type: SubscriptionType.Author,
 			id: authorId,
-			isAuthorized: true,
-			state: 'teststateofauthor'
+			isAuthorized: true
 		};
 		const isAuthor = true;
 		jest.spyOn(subscriptionService, 'isAuthor').mockImplementation(async () => isAuthor);
+		jest.spyOn(subscriptionService, 'getSubscriptionState').mockImplementationOnce(async () => 'teststateofauthor');
 		jest
 			.spyOn(subscriptionService, 'getSubscription')
 			.mockImplementationOnce(async () => subscriptionMock)
@@ -179,11 +177,11 @@ describe('test authorize subscription route', () => {
 			type: SubscriptionType.Author,
 			id: authorId,
 			isAuthorized: true,
-			state: 'teststateofauthor',
 			pskId
 		};
 		const isAuthor = true;
 		jest.spyOn(subscriptionService, 'isAuthor').mockImplementation(async () => isAuthor);
+		jest.spyOn(subscriptionService, 'getSubscriptionState').mockImplementationOnce(async () => 'teststateofauthor');
 		jest
 			.spyOn(subscriptionService, 'getSubscription')
 			.mockImplementationOnce(async () => subscriptionMock)
@@ -220,18 +218,20 @@ describe('test authorize subscription route', () => {
 			type: SubscriptionType.Author,
 			id: authorId,
 			isAuthorized: true,
-			state: 'teststateofauthor',
 			pskId // presharedKey is undefined
 		};
 		const isAuthor = true;
 		jest.spyOn(subscriptionService, 'isAuthor').mockImplementation(async () => isAuthor);
+		jest.spyOn(subscriptionService, 'getSubscriptionState').mockImplementationOnce(async () => 'teststateofauthor');
 		jest
 			.spyOn(subscriptionService, 'getSubscription')
 			.mockImplementationOnce(async () => subscriptionMock)
 			.mockImplementationOnce(async () => author);
 		jest.spyOn(subscriptionDb, 'getSubscriptions').mockImplementation(async () => []);
 		const authorMock = AuthorMock;
-		const removeChannelRequestedSubscriptionIdSpy = jest.spyOn(channelInfoService, 'removeChannelRequestedSubscriptionId').mockImplementation(async () => null);
+		const removeChannelRequestedSubscriptionIdSpy = jest
+			.spyOn(channelInfoService, 'removeChannelRequestedSubscriptionId')
+			.mockImplementation(async () => null);
 		const addChannelSubscriberIdSpy = jest.spyOn(channelInfoService, 'addChannelSubscriberId').mockImplementation(async () => null);
 		const receiveSubscribeSpy = jest.spyOn(streamsService, 'receiveSubscribe').mockImplementation(async () => null);
 		const importAuthorSpy = jest.spyOn(streamsService, 'importSubscription').mockImplementation(async () => authorMock); // author found
@@ -257,8 +257,8 @@ describe('test authorize subscription route', () => {
 		expect(setSubscriptionAuthorizedSpy).toHaveBeenCalledWith('testaddress', authorId, 'testkeyloadlink', 'testsequencelink');
 		expect(exportSubscriptionSpy).toHaveBeenCalledWith(authorMock, 'veryvery-very-very-server-secret');
 		expect(updateSubscriptionStateSpy).toHaveBeenCalledWith('testaddress', authorId, 'new-state');
-		expect(removeChannelRequestedSubscriptionIdSpy).toHaveBeenCalledWith('testaddress','did:iota:1234');
-		expect(addChannelSubscriberIdSpy).toHaveBeenCalledWith('testaddress','did:iota:1234');
+		expect(removeChannelRequestedSubscriptionIdSpy).toHaveBeenCalledWith('testaddress', 'did:iota:1234');
+		expect(addChannelSubscriberIdSpy).toHaveBeenCalledWith('testaddress', 'did:iota:1234');
 		expect(res.status).toHaveBeenCalledWith(StatusCodes.OK);
 		expect(res.send).toHaveBeenCalledWith({ keyloadLink: 'testkeyloadlink', sequenceLink: 'testsequencelink' });
 	});
@@ -272,11 +272,11 @@ describe('test authorize subscription route', () => {
 			type: SubscriptionType.Author,
 			id: authorId,
 			isAuthorized: true,
-			state: 'teststateofauthor',
 			pskId
 		};
 		const isAuthor = true;
 		jest.spyOn(subscriptionService, 'isAuthor').mockImplementation(async () => isAuthor);
+		jest.spyOn(subscriptionService, 'getSubscriptionState').mockImplementationOnce(async () => 'teststateofauthor');
 		jest
 			.spyOn(subscriptionService, 'getSubscription')
 			.mockImplementationOnce(async () => subscriptionMock)
@@ -284,7 +284,9 @@ describe('test authorize subscription route', () => {
 
 		jest.spyOn(subscriptionDb, 'getSubscriptions').mockImplementation(async () => []);
 		const authorMock = AuthorMock;
-		const removeChannelRequestedSubscriptionIdSpy = jest.spyOn(channelInfoService, 'removeChannelRequestedSubscriptionId').mockImplementation(async () => null);
+		const removeChannelRequestedSubscriptionIdSpy = jest
+			.spyOn(channelInfoService, 'removeChannelRequestedSubscriptionId')
+			.mockImplementation(async () => null);
 		const addChannelSubscriberIdSpy = jest.spyOn(channelInfoService, 'addChannelSubscriberId').mockImplementation(async () => null);
 		const receiveSubscribeSpy = jest.spyOn(streamsService, 'receiveSubscribe').mockImplementation(async () => null);
 		const importAuthorSpy = jest.spyOn(streamsService, 'importSubscription').mockImplementation(async () => authorMock); // author found
@@ -310,8 +312,8 @@ describe('test authorize subscription route', () => {
 		expect(setSubscriptionAuthorizedSpy).toHaveBeenCalledWith('testaddress', authorId, 'testkeyloadlink', 'testsequencelink');
 		expect(exportSubscriptionSpy).toHaveBeenCalledWith(authorMock, 'veryvery-very-very-server-secret');
 		expect(updateSubscriptionStateSpy).toHaveBeenCalledWith('testaddress', authorId, 'new-state');
-		expect(removeChannelRequestedSubscriptionIdSpy).toHaveBeenCalledWith('testaddress','did:iota:1234');
-		expect(addChannelSubscriberIdSpy).toHaveBeenCalledWith('testaddress','did:iota:1234');
+		expect(removeChannelRequestedSubscriptionIdSpy).toHaveBeenCalledWith('testaddress', 'did:iota:1234');
+		expect(addChannelSubscriberIdSpy).toHaveBeenCalledWith('testaddress', 'did:iota:1234');
 		expect(res.status).toHaveBeenCalledWith(StatusCodes.OK);
 		expect(res.send).toHaveBeenCalledWith({ keyloadLink: 'testkeyloadlink', sequenceLink: 'testsequencelink' });
 	});