+++ api/shared-modules/src/web/models/schemas/index.ts
@@ -14,9 +14,10 @@ export {
 	IdentityKeysSchema,
 	IdentityKeyPairSchema,
 	VerifiableCredentialSchema,
+	VerifiablePresentationSchema,
 	VerifiableCredentialSubjectSchema
 } from './identity';
-export { AccessRights, SubscriptionSchema, SubscriptionType, SubscriptionUpdateSchema } from './subscription';
+export { AccessRights, SubscriptionSchema, SubscriptionType, SubscriptionStateSchema, SubscriptionUpdateSchema } from './subscription';
 export {
 	AggregateOfferSchema,
 	AggregateRatingSchema,