+++ api/shared-modules/src/web/models/schemas/subscription.ts
@@ -16,7 +16,6 @@ export const SubscriptionSchema = Type.Object({
 	type: Type.Enum(SubscriptionType),
 	channelAddress: Type.String({ minLength: 105, maxLength: 105 }),
 	id: Type.String({ minLength: 50, maxLength: 53 }),
-	state: Type.String(),
 	subscriptionLink: Type.Optional(Type.String()),
 	isAuthorized: Type.Boolean(),
 	accessRights: Type.Enum(AccessRights),
@@ -26,11 +25,14 @@ export const SubscriptionSchema = Type.Object({
 	pskId: Type.Optional(Type.String())
 });
 
+export const SubscriptionStateSchema = Type.Object({
+	state: Type.String()
+});
+
 export const SubscriptionUpdateSchema = Type.Object({
 	type: Type.Optional(Type.Enum(SubscriptionType)),
 	channelAddress: Type.Optional(Type.String({ minLength: 105, maxLength: 105 })),
 	id: Type.Optional(Type.String({ minLength: 50, maxLength: 53 })),
-	state: Type.Optional(Type.String()),
 	subscriptionLink: Type.Optional(Type.String()),
 	isAuthorized: Type.Optional(Type.Boolean()),
 	accessRights: Type.Optional(Type.Enum(AccessRights)),