+++ api/shared-modules/src/web/models/types/index.ts
@@ -11,6 +11,7 @@ export {
 	IdentitySearchBody,
 	LatestIdentity,
 	VerifiableCredential,
+	VerifiablePresentation,
 	VerifiableCredentialPersistence,
 	Bitmap,
 	IdentityKeys,
@@ -35,10 +36,9 @@ export {
 	TrustedRootBody,
 	ValidateBody,
 	ValidateResponse,
-	VerifiableCredentialBody,
 	VerifyJwtBody
 } from './request-response-bodies';
-export { Subscription, SubscriptionUpdate } from './subscription';
+export { Subscription, SubscriptionState, SubscriptionUpdate } from './subscription';
 export {
 	Device,
 	IdentityClaim,
@@ -53,4 +53,4 @@ export {
 	UserSearchResponse,
 	UserType
 } from './user';
-export { AuthenticatedRequest, AuthorizationCheck, CredentialTypes, Subject, VerifiableCredentialInternal } from './verification';
+export { AuthenticatedRequest, AuthorizationCheck, CredentialTypes, Subject } from './verification';