+++ clients/client-sdk/src/clients/identity.ts
@@ -7,12 +7,22 @@ import {
   User,
   UserType,
   CredentialTypes,
-  VerifiableCredentialInternal,
   IdentityDocument,
   IdentityKeys,
   UserSearchResponse
 } from '@iota/is-shared-modules';
 import { SearchCriteria } from '../models/searchCriteria';
+import * as bs58 from 'bs58';
+import {
+  Credential,
+  Duration,
+  Document,
+  KeyPair,
+  Presentation,
+  Timestamp,
+  IPresentation,
+  ProofOptions
+} from '@iota/identity-wasm/web';
 
 export class IdentityClient extends BaseClient {
   private baseUrl: string;
@@ -112,6 +122,85 @@ export class IdentityClient extends BaseClient {
     });
   }
 
+  private async restoreIdentity(
+    identity: IdentityKeys
+  ): Promise<{ document: Document; key: KeyPair; messageId: string }> {
+    try {
+      const decodedKey = {
+        public: Array.from(bs58.decode(identity.keys.sign.public)),
+        secret: Array.from(bs58.decode(identity.keys.sign.private))
+      };
+      const json = {
+        type: identity.keys.sign.type,
+        public: decodedKey.public,
+        private: decodedKey.secret
+      };
+      const key: KeyPair = KeyPair.fromJSON(json);
+      const { document, messageId } = await this.latestDocument(identity.id);
+      return {
+        document: Document.fromJSON(document),
+        key,
+        messageId
+      };
+    } catch (error) {
+      console.error(`Error from identity sdk: ${error}`);
+      throw new Error('could not parse key or doc of the identity');
+    }
+  }
+  /**
+   * Create a Verifiable Presentation.
+   *
+   * @param {{
+   *     signedVcJson: any;
+   *     identityKeys: IdentityKeys;
+   *     challenge?: string;
+   *     expiration?: number;
+   *   }} props Properties
+   * @param signedVcJson: Signed Verifiable Credential.
+   * @param identityKeys: Identity keys to sign the Verifiable Presentation.
+   * @param challenge: Challenge to mitigate replay attacks.
+   * @param expiration: Time when the presentation shall expire in seconds.
+   * @return {*}  {Promise<void>}
+   * @memberof IdentityClient
+   */
+  async createVerifiablePresentation(props: {
+    signedVcJson: any | any[];
+    identityKeys: IdentityKeys;
+    challenge?: string;
+    expiration?: number;
+  }): Promise<void> {
+    const { signedVcJson, identityKeys, challenge, expiration } = props;
+    const expires =
+      expiration != null ? Timestamp.nowUTC().checkedAdd(Duration.seconds(expiration)) : undefined;
+    const identity = await this.restoreIdentity(identityKeys);
+
+    let receivedVC: Credential | Credential[];
+    if (Array.isArray(signedVcJson)) {
+      receivedVC = signedVcJson.map((vc) => Credential.fromJSON(vc));
+    } else {
+      receivedVC = Credential.fromJSON(signedVcJson);
+    }
+
+    const pres: IPresentation = {
+      verifiableCredential: receivedVC,
+      holder: identity.document.id()
+    };
+
+    const unsignedVp = new Presentation(pres);
+    const methodId = identity.document.defaultSigningMethod().id().toString();
+    const signedVp = await identity.document.signPresentation(
+      unsignedVp,
+      identity.key.private(),
+      methodId,
+      new ProofOptions({
+        challenge,
+        expires
+      })
+    );
+
+    return signedVp.toJSON();
+  }
+
   /**
    * Get the latest version of an identity document (DID) from the IOTA Tangle.
    * @param id
@@ -158,7 +247,7 @@ export class IdentityClient extends BaseClient {
    * @returns
    */
   async createCredential(
-    initiatorVC: VerifiableCredentialInternal | undefined,
+    initiatorVC: VerifiableCredential | undefined,
     targetDid: string,
     credentialType: CredentialTypes | string,
     claimType: UserType,